+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
+## [1.1.0][] - 2019-10-11
+
 ### Added
 - Guild summary on home page
 - WvW Score page
@@ -18,5 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 
-[Unreleased]: https://github.com/Nabrok/gw2-discord-spa/compare/v1.0.0...master
+[Unreleased]: https://github.com/Nabrok/gw2-discord-spa/compare/v1.1.0...master
+[1.1.0]: https://github.com/Nabrok/gw2-discord-spa/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Nabrok/gw2-discord-spa/tags/v1.0.0